requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
imageio==2.29.0
importlib-metadata==6.6.0
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.6.3
ninja==1.11.1
numpy==1.21.6
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
psutil==5.9.5
pyclipper==1.3.0.post4
pyparsing==3.0.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.30.0
scikit-image==0.19.3
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.23.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.15
thop==0.1.1.post2209072238
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
ultralytics==8.0.105
urllib3==1.26.15
zipp==3.15.0
